+++ test/code-types/ExtractAstTest.js
@@ -2,6 +2,7 @@
 Object.defineProperty(exports, "__esModule", { value: true });
 var chai = require("chai");
 var ExtractAst = require("../../code-types/ExtractAst");
+var TypeConverter = require("../../code-types/TypeConverter");
 var asr = chai.assert;
 suite("ExtractAst", function ExtractAstTest() {
     function getModelAsts() {
@@ -63,9 +64,9 @@ suite("ExtractAst", function ExtractAstTest() {
                 "blockType": "CLASS",
                 "using": ["System.Collections.Generic", "System.Runtime.Serialization"],
                 "fields": [
-                    { "name": "PageNumber", "type": { "typeName": "int", "primitive": true }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>The page number.</value>"] },
-                    { "name": "TotalRecords", "type": { "typeName": "int", "primitive": true }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>The total record count.</value>"] },
-                    { "name": "Items", "type": { "typeName": "IList", "genericParameters": [{ "typeName": "T" }] }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>The search result items.</value>"] }
+                    { "name": "PageNumber", "type": { "typeName": "int", "primitive": true }, "accessModifiers": ["public"], "annotations": [], "comments": [" <value>The page number.</value>"] },
+                    { "name": "TotalRecords", "type": { "typeName": "int", "primitive": true }, "accessModifiers": ["public"], "annotations": [], "comments": [" <value>The total record count.</value>"] },
+                    { "name": "Items", "type": { "typeName": "IList", "genericParameters": [{ "typeName": "T" }] }, "accessModifiers": ["public"], "annotations": [], "comments": [" <value>The search result items.</value>"] }
                 ],
                 "methods": [
                     { "name": "ToString", "parameters": [], "accessModifiers": [], "annotations": [], "returnType": { "typeName": "string" }, "comments": [] }
@@ -73,26 +74,26 @@ suite("ExtractAst", function ExtractAstTest() {
             },
             // Search criteria
             {
-                "classSignature": { "access": "PUBLIC", "name": "TestApp.Models.Base.SearchCriteria", "declarationType": "class", "annotations": [{ "name": "DataContract", "arguments": {} }] },
+                "classSignature": { "access": "PUBLIC", "name": "TestApp.Models.Base.SearchCriteria", "declarationType": "class", "annotations": [{ "name": "Criteria", "arguments": { "value": "TestApp.Models.Base.SearchCriteria" } }] },
                 "blockType": "CLASS",
                 "using": ["System.Runtime.Serialization", "TestApp.Models.Enums"],
                 "fields": [
-                    { "name": "PageSize", "type": { "typeName": "int", "primitive": true }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>The page size. Number of records.</value>"] },
-                    { "name": "PageNumber", "type": { "typeName": "int", "primitive": true }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>The page number. If PageNumber is present, PageSize must be present as well. Multiply by 'PageSize' to get the record offset</value>"] },
-                    { "name": "SortBy", "type": { "typeName": "string" }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>The column name to sort by.</value>"] },
-                    { "name": "SortOrder", "type": { "typeName": "TestApp.Models.Enums.SortOrder" }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [] }
+                    { "name": "PageSize", "type": { "typeName": "int", "primitive": true }, "accessModifiers": ["public"], "annotations": [{ "name": "Criteria", "arguments": {} }], "comments": [" <value>The page size. Number of records.</value>"] },
+                    { "name": "PageNumber", "type": { "typeName": "int", "primitive": true }, "accessModifiers": ["public"], "annotations": [{ "name": "Criteria", "arguments": {} }], "comments": [" <value>The page number. If PageNumber is present, PageSize must be present as well. Multiply by 'PageSize' to get the record offset</value>"] },
+                    { "name": "SortBy", "type": { "typeName": "string" }, "accessModifiers": ["public"], "annotations": [{ "name": "Criteria", "arguments": {} }], "comments": [" <value>The column name to sort by.</value>"] },
+                    { "name": "SortOrder", "type": { "typeName": "TestApp.Models.Enums.SortOrder" }, "accessModifiers": ["public"], "annotations": [{ "name": "Criteria", "arguments": { "JsonType": "int" } }], "comments": [] }
                 ],
                 "methods": []
             }, {
-                "classSignature": { "access": "PUBLIC", "name": "TestApp.Models.Searching.CoverLogSearchCriteria", "declarationType": "class", "extendClassName": { "typeName": "TestApp.Models.Base.SearchCriteria" }, "annotations": [{ "name": "DataContract", "arguments": {} }] },
+                "classSignature": { "access": "PUBLIC", "name": "TestApp.Models.Searching.CoverLogSearchCriteria", "declarationType": "class", "extendClassName": { "typeName": "TestApp.Models.Base.SearchCriteria" }, "annotations": [{ "name": "Criteria", "arguments": { "value": "string" } }] },
                 "blockType": "CLASS",
                 "using": ["System", "System.Runtime.Serialization", "TestApp.Models.Base"],
                 "fields": [
-                    { "name": "ServiceNames", "type": { "typeName": "string", "arrayDimensions": 1 }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>Service names to search for.</value>"] },
-                    { "name": "MethodNames", "type": { "typeName": "string", "arrayDimensions": 1 }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>Method names to search for.</value>"] },
-                    { "name": "ServiceMethodNames", "type": { "typeName": "string", "arrayDimensions": 1 }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>Compound 'service.menthod' format names to search for.</value>"] },
-                    { "name": "SearchPeriodStart", "type": { "typeName": "DateTime" }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>The start of the search period.</value>"] },
-                    { "name": "SearchPeriodEnd", "type": { "typeName": "DateTime" }, "accessModifiers": ["public"], "annotations": [{ "name": "DataMember", "arguments": {} }], "comments": [" <value>The end of the search period.</value>"] }
+                    { "name": "ServiceNames", "type": { "typeName": "string", "arrayDimensions": 1 }, "accessModifiers": ["public"], "annotations": [{ "name": "Criteria", "arguments": { "JsonType": "string[]" } }], "comments": [" <value>Service names to search for.</value>"] },
+                    { "name": "MethodNames", "type": { "typeName": "string", "arrayDimensions": 1 }, "accessModifiers": ["public"], "annotations": [{ "name": "Criteria", "arguments": { "JsonType": "string[]" } }], "comments": [" <value>Method names to search for.</value>"] },
+                    { "name": "ServiceMethodNames", "type": { "typeName": "string", "arrayDimensions": 1 }, "accessModifiers": ["public"], "annotations": [{ "name": "Criteria", "arguments": { "JsonType": "string[]" } }], "comments": [" <value>Compound 'service.menthod' format names to search for.</value>"] },
+                    { "name": "SearchPeriodStart", "type": { "typeName": "DateTime" }, "accessModifiers": ["public"], "annotations": [{ "name": "Criteria", "arguments": { "JsonType": "string" } }], "comments": [" <value>The start of the search period.</value>"] },
+                    { "name": "SearchPeriodEnd", "type": { "typeName": "DateTime" }, "accessModifiers": ["public"], "annotations": [{ "name": "Criteria", "arguments": { "JsonType": "string" } }], "comments": [" <value>The end of the search period.</value>"] }
                 ],
                 "methods": []
             }];
@@ -201,6 +202,19 @@ suite("ExtractAst", function ExtractAstTest() {
         };
         asr.deepEqual(Object.keys(res).sort(), ["IList", "T", "TestApp.Models.Enums.SortOrder", "TestApp.Models.PhotoHeader", "UserIdentifier"]);
         //console.log(JSON.stringify(res));
+    });
+    test("extractAnnotationArgumentTypes", function extractAnnotationArgumentTypesTest() {
+        var models = toMap(getModelAsts(), function (obj) { return obj.classSignature.name; });
+        var res = ExtractAst.extractAnnotationArgumentTypes(Object.keys(models), models, true, function (annot, argName, argValue) {
+            console.log("annotation '" + argName + "': " + argValue);
+            var type = TypeConverter.parseTypeTemplate(argValue).typeName;
+            return (TypeConverter.isPrimitive(type) || TypeConverter.isCore(type) || models[type] != null) ? argValue : null;
+        });
+        asr.deepEqual(Object.keys(res).map(function (k) { return res[k].map(function (au) { return au.annotation; }); }), [[{
+                    "name": "Criteria", "arguments": { "value": "TestApp.Models.Base.SearchCriteria" }
+                }], [{
+                    "name": "Criteria", "arguments": { "value": "string" }
+                }]]);
     });
     function toMap(ary, keyGetter) {
         var map = {};